Master's Thesis Code
This is the code I have written for my Master's thesis: continuous time scale modeling in psychopathological networks. The code consists of two major parts: a simulation analysis and an empirical analysis. However, fundamentally, it is the same process. A network model with no latent specifications is constructed using the ctmodel function. Essentially, this is a continuous time model using first-order stochastic differential equations. This model can be used to do various things as seen in the algorithm (i.e., simulate data, fit model to data and estimate parameters).